refactor: user/pay 도메인 삭제
