+++ app/src/main/java/com/joonhyeok/app/user/application/OutboxService.java
@@ -1,9 +1,9 @@
 package com.joonhyeok.app.user.application;
 
-import com.joonhyeok.app.user.application.outbox.OutboxFindCommand;
-import com.joonhyeok.app.user.application.outbox.OutboxSaveCommand;
-import com.joonhyeok.app.user.application.outbox.OutboxSendFailCommand;
-import com.joonhyeok.app.user.application.outbox.OutboxSendSuccessCommand;
+import com.joonhyeok.app.user.application.dto.outbox.OutboxFindCommand;
+import com.joonhyeok.app.user.application.dto.outbox.OutboxSaveCommand;
+import com.joonhyeok.app.user.application.dto.outbox.OutboxSendFailCommand;
+import com.joonhyeok.app.user.application.dto.outbox.OutboxSendSuccessCommand;
 import com.joonhyeok.app.user.domain.outbox.Outbox;
 import com.joonhyeok.app.user.domain.outbox.OutboxRepository;
 import jakarta.persistence.EntityNotFoundException;
