+++ app/src/main/java/com/joonhyeok/app/user/application/PayInfoService.java
@@ -1,6 +1,6 @@
 package com.joonhyeok.app.user.application;
 
-import com.joonhyeok.app.user.domain.pay.PayEvent;
+import com.joonhyeok.app.user.domain.user.PayEvent;
 
 public interface PayInfoService {
      void sendPayInfo(PayEvent payEvent);
