+++ app/src/main/java/com/joonhyeok/app/user/application/UserPayService.java
@@ -6,10 +6,10 @@ import com.joonhyeok.app.queue.domain.Queue;
 import com.joonhyeok.app.queue.domain.QueueRepository;
 import com.joonhyeok.app.reservation.domain.Reservation;
 import com.joonhyeok.app.reservation.domain.ReservationRepository;
-import com.joonhyeok.app.user.application.dto.UserPayCommand;
-import com.joonhyeok.app.user.application.dto.UserPayResult;
-import com.joonhyeok.app.user.domain.pay.PayEvent;
-import com.joonhyeok.app.user.domain.pay.PayEventPublisher;
+import com.joonhyeok.app.user.application.dto.user.UserPayCommand;
+import com.joonhyeok.app.user.application.dto.user.UserPayResult;
+import com.joonhyeok.app.user.domain.user.PayEvent;
+import com.joonhyeok.app.user.domain.user.PayEventPublisher;
 import com.joonhyeok.app.user.domain.user.PayValidator;
 import com.joonhyeok.app.user.domain.user.User;
 import com.joonhyeok.app.user.domain.user.UserRepository;
