+++ app/src/main/java/com/joonhyeok/app/user/application/dto/outbox/OutboxSendFailCommand.java
@@ -1,4 +1,4 @@
-package com.joonhyeok.app.user.application.outbox;
+package com.joonhyeok.app.user.application.dto.outbox;
 
 public record OutboxSendFailCommand(
         String type,
