+++ app/src/main/java/com/joonhyeok/app/user/application/dto/user/AccountBalanceQueryResult.java
@@ -1,4 +1,4 @@
-package com.joonhyeok.app.user.application.dto;
+package com.joonhyeok.app.user.application.dto.user;
 
 public record AccountBalanceQueryResult(
         Long balance
