+++ app/src/main/java/com/joonhyeok/app/user/application/dto/user/UserPayCommand.java
@@ -1,4 +1,4 @@
-package com.joonhyeok.app.user.application.dto;
+package com.joonhyeok.app.user.application.dto.user;
 
 public record UserPayCommand(
         Long userId,
