+++ app/src/main/java/com/joonhyeok/app/user/application/dto/user/UserPayResult.java
@@ -1,4 +1,4 @@
-package com.joonhyeok.app.user.application.dto;
+package com.joonhyeok.app.user.application.dto.user;
 
 public record UserPayResult(
         Long reservationId
