+++ app/src/main/java/com/joonhyeok/app/user/domain/outbox/PayOutboxListener.java
@@ -1,6 +1,6 @@
 package com.joonhyeok.app.user.domain.outbox;
 
-import com.joonhyeok.app.user.domain.pay.PayEvent;
+import com.joonhyeok.app.user.domain.user.PayEvent;
 
 public interface PayOutboxListener {
     public void listen(PayEvent payEvent);
