--- app/src/main/java/com/joonhyeok/app/user/domain/pay/PayEvent.java
@@ -1,8 +1,0 @@
-package com.joonhyeok.app.user.domain.pay;
-
-import com.joonhyeok.app.user.application.dto.UserPayResult;
-
-public record PayEvent(
-        UserPayResult result
-) {
-}
