+++ app/src/main/java/com/joonhyeok/app/user/domain/user/PayEvent.java
@@ -1,0 +1,8 @@
+package com.joonhyeok.app.user.domain.user;
+
+import com.joonhyeok.app.user.application.dto.user.UserPayResult;
+
+public record PayEvent(
+        UserPayResult result
+) {
+}
