+++ app/src/main/java/com/joonhyeok/app/user/domain/user/PayEventListener.java
@@ -1,4 +1,4 @@
-package com.joonhyeok.app.user.domain.pay;
+package com.joonhyeok.app.user.domain.user;
 
 public interface PayEventListener {
     public void sendPayInfo(PayEvent payEvent);
