+++ app/src/main/java/com/joonhyeok/app/user/domain/user/PayEventPublisher.java
@@ -1,4 +1,4 @@
-package com.joonhyeok.app.user.domain.pay;
+package com.joonhyeok.app.user.domain.user;
 
 public interface PayEventPublisher {
     public void publish(PayEvent payEvent);
