+++ app/src/main/java/com/joonhyeok/app/user/infra/domain/outbox/PayOutboxKafkaListener.java
@@ -1,8 +1,8 @@
 package com.joonhyeok.app.user.infra.domain.outbox;
 
 import com.joonhyeok.app.user.domain.outbox.PayOutboxListener;
-import com.joonhyeok.app.user.domain.pay.PayEvent;
-import com.joonhyeok.app.user.application.outbox.OutboxSendSuccessCommand;
+import com.joonhyeok.app.user.domain.user.PayEvent;
+import com.joonhyeok.app.user.application.dto.outbox.OutboxSendSuccessCommand;
 import com.joonhyeok.app.user.application.OutboxService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
