+++ app/src/test/java/com/joonhyeok/app/user/DummyEventPublisher.java
@@ -1,7 +1,7 @@
 package com.joonhyeok.app.user;
 
-import com.joonhyeok.app.user.domain.pay.PayEvent;
-import com.joonhyeok.app.user.domain.pay.PayEventPublisher;
+import com.joonhyeok.app.user.domain.user.PayEvent;
+import com.joonhyeok.app.user.domain.user.PayEventPublisher;
 
 public class DummyEventPublisher implements PayEventPublisher {
